Gender-Age-Detection/requirements.txt
opencv-python>=4.5.0
numpy>=1.19.0
streamlit>=1.0.0
Pillow>=8.0.0
argparse>=1.4.0
gdown>=4.7.1
requests>=2.31.0 